requirements.txt
ta==0.10.2
pandas==2.0.2
python-binance==1.0.17
pipdeptree==2.9.3
pyaml==23.5.9
requests==2.31.0
fastparquet==2023.4.0
scikit-learn==1.2.2